Move SWTMessageBoxType from model into view
It purely concerns the view.  I guess it was in model because it is abstract and does not have code that actually modifies what the user sees, but it clearly is only meaningful within the view.
